requirements.txt
PyPDF2>=3.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pymupdf>=1.23.0
imagehash>=4.3.0
Pillow>=10.0.0
